Move the quarantine code from rest to handlers.
This makes it easier to override in downstream plugins.
