+++ histomicsui/rest/hui_resource.py
@@ -1,89 +1,16 @@
-import datetime
-
 from girder import logger
 from girder.api import access
 from girder.api.rest import Resource, filtermodel
 from girder.api.describe import autoDescribeRoute, describeRoute, Description
 from girder.constants import AccessType, TokenScope
 from girder.exceptions import RestException
-from girder.models.folder import Folder
 from girder.models.item import Item
 from girder.models.setting import Setting
 from girder.utility.model_importer import ModelImporter
 
 from ..constants import PluginSettings
+from .. import handlers
 from .system import allChildFolders, allChildItems
-
-
-def quarantine_item(item, user, makePlaceholder=True):
-    """
-    Quarantine an item, marking which user did it.  Note that this raises
-    RestExceptions for failures.
-
-    :param user: the user doing the quarantining.
-    :param item: an item to quarantine.
-    :returns: the modified item.
-    """
-    folder = Setting().get(PluginSettings.HUI_QUARANTINE_FOLDER)
-    if not folder:
-        raise RestException('The quarantine folder is not configured.')
-    folder = Folder().load(folder, force=True, exc=True)
-    if not folder:
-        raise RestException('The quarantine folder does not exist.')
-    if str(folder['_id']) == str(item['folderId']):
-        raise RestException('The item is already in the quarantine folder.')
-    originalFolder = Folder().load(item['folderId'], force=True)
-    quarantineInfo = {
-        'originalFolderId': item['folderId'],
-        'originalBaseParentType': item['baseParentType'],
-        'originalBaseParentId': item['baseParentId'],
-        'originalUpdated': item['updated'],
-        'quarantineUserId': user['_id'],
-        'quarantineTime': datetime.datetime.utcnow()
-    }
-    item = Item().move(item, folder)
-    if makePlaceholder:
-        placeholder = Item().createItem(
-            item['name'] + ' [Removed - Quarantined]',
-            {'_id': item['creatorId']}, originalFolder,
-            description=item['description'])
-        quarantineInfo['placeholderItemId'] = placeholder['_id']
-    item.setdefault('meta', {})['quarantine'] = quarantineInfo
-    item = Item().updateItem(item)
-    if makePlaceholder:
-        placeholderInfo = {
-            'quarantined': True,
-            'quarantineTime': quarantineInfo['quarantineTime']
-        }
-        placeholder.setdefault('meta', {})['quarantine'] = placeholderInfo
-        placeholder = Item().updateItem(placeholder)
-    return item
-
-
-def restore_quarantine_item(item):
-    """
-    Unquarantine an item, returning it to its original location.  Note that
-    this raises RestExceptions for failures.
-
-    :param item: an item to unquarantine.
-    :returns: the modified item.
-    """
-    if not item.get('meta', {}).get('quarantine'):
-        raise RestException('The item has no quarantine record.')
-    folder = Folder().load(item['meta']['quarantine']['originalFolderId'], force=True)
-    if not folder:
-        raise RestException('The original folder is not accessible.')
-    if 'placeholderItemId' in item['meta']['quarantine']:
-        placeholder = Item().load(item['meta']['quarantine']['placeholderItemId'], force=True)
-    else:
-        placeholder = None
-    item = Item().move(item, folder)
-    item['updated'] = item['meta']['quarantine']['originalUpdated']
-    del item['meta']['quarantine']
-    item = Item().updateItem(item)
-    if placeholder is not None:
-        Item().remove(placeholder)
-    return item
 
 
 class HistomicsUIResource(Resource):
@@ -138,7 +65,7 @@ class HistomicsUIResource(Resource):
     @access.user(scope=TokenScope.DATA_WRITE)
     @filtermodel(model=Item)
     def putQuarantine(self, item):
-        return quarantine_item(item, self.getCurrentUser())
+        return handlers.quarantine_item(item, self.getCurrentUser())
 
     @autoDescribeRoute(
         Description('Restore a quarantined item to its original folder.')
@@ -150,7 +77,7 @@ class HistomicsUIResource(Resource):
     @access.admin
     @filtermodel(model=Item)
     def restoreQuarantine(self, item):
-        return restore_quarantine_item(item)
+        return handlers.restore_quarantine_item(item, self.getCurrentUser())
 
     # The `autoDescrbeRoute` (and `describeRoute` used in older code)
     # serves to generate the swagger documentation that looks like:
